had to revert move of insertVariable
This was failing because the editor saved in the plugin
by the onInitEditor was not working for the insert.
Even if I changed it so the editor was set after all of the
initialization it still didn't work correctly. :shrug:

Moving those functions into the class weren't very useful anyhow.