fix functions build and deploy
The previous build was putting the built files in `lib/src/` instead
of just `lib/`. This is fixed by adding a tsconfig.prod.json

Also randomly the build started failing with a typescript error.
The fix for this was to bump the version of @google-cloud/common

Additionally the deployed zip file for each function was including
the uncompiled src and test folders,
so an `ignore` was added to firebase.json
The default `ignore` list included several things which are not included
when `ignore` is explicitly specified, so those items are added too.