fix: New variable dialog creates two undoable actions (PT-184406182)
PT Story: https://www.pivotaltracker.com/story/show/184406182

Adds a `noUndo` option to SharedVariable's `addAndInsertVariable` action. The new option is used to determine if `withoutUndo` should be called for the action.

In some cases (e.g., adding a new variable to a text tile), using the `addAndInsertVariable` action can result in two undoable actions being added to the undo history which can lead to confusing and unexpected behavior. Passing `true` for the `noUndo` option ensures that only one undo action is added to the undo history, preventing the unexpected behavior.