fix: Changes to variable cards not saved unless click away (PT-184474083)
PT Story: https://www.pivotaltracker.com/story/show/184474083

These changes fix an issue due to the use of useMemo for creating `variableClone`. The source `variable` was included in the call to useMemo's dependency array, but because useMemo only does a shallow comparison of objects, changes to `variable`'s properties were not triggering the re-creation of `variableClone`.