fix: Initial load from production portal using qa log ingester (PT-184068699)
PT Story: https://www.pivotaltracker.com/story/show/184068699

The `portal` property of user is initially empty when the app first loads, so `sendToLoggingService` was defaulting to using the QA URL. The `portal` property in `data.parameters` seems to contain the expected value, so these changes include an additional check for that for when `user.portal` is empty.

Included is a change to the `console.log` statement when `DEBUG_LOGGER` is true. We were previously outputting the full `logManagerUrl` object instead of the specific log ingester URL being used. It seems like we would want the specific URL, but I want to double check that.