fix: Table and Graph do not synchronize in 4-up view (PT-184609260)
PT Story: https://www.pivotaltracker.com/story/show/184609260

While a teacher is viewing a student's work in the 4-up view, if the student links a graph and a table, the data from the table does not appear in the graph in the teacher's view (it does appear for the student). If the teacher reloads the page and navigates back to the view of the student's work, the graph is then populated.

When the link is first made, an error appears in the teacher's browser console: `Uncaught (in promise) TypeError: Cannot read properties of undefined (reading 'fill')`. These changes prevent that error by checking if `colorMapEntry.colorset` is defined instead of just checking for `colorMapEntry` since `colorMapEntry` can exist with an undefined `colorSet`. With that error prevented, the graph is populated immediately in the teacher's view.