use a class component for the custom widget (#1631)
* use a class component for the custom widget

I was seeing an error in the console about one of the props passed to the
function component. And also the Decap docs imply the components
should be class components.

Probably more important that this change is the use of `toJS()`. It seems
the CMS is using redux with immutable (or something like it). Because of
this, it converts all objects to be "immutable" objects. Those types of
objects can be converted back into plain javascript objects using
`toJS()`.  I came up with that idea because I saw lots of references to
it in the Decap source code.

* Pretty loaded json. Switch app.sass to scss. Set up admin.html to make links work in cms when using ?admin

* Turn off adding sections through cms

---------

Co-authored-by: Teale Fristoe <fristoe@gmail.com>