extract document monitoring to its own class
This uses a MobX autorun to monitor both the documents and groups
This approach consolidates the logic for deciding which documents
to monitor into one place instead of having it in both the
problem-documents-listener and groups-listener.

It also reduces the firebase listener churn that happening before.
Listeners would be added and then removed and then added again.