fix: Styling missing on links (PT-184882521)
PT Story: https://www.pivotaltracker.com/story/show/184882521

In the CMS, links are rendered in dark gray text and do not have an underline, which makes them difficult to see. These changes revert the color and text-decoration CSS rules for links within the CLUE canvas part of the CMS so links within tiles are more visible.