docs: Use iframe for CMS
PT Story: https://www.pivotaltracker.com/story/show/184919981

We were originally going to automatically install the CMS dependencies by adding `&& cd cms && npm install` to the `postinstall` script in the root package.json. That wouldn't work in the continuous integration process on GitHub, so we had to abandon that. Unfortunately, this means developers and other users who want to use CLUE locally will need to take the additional step noted in the readme under "Initial steps." I neglected to adjust the readme after we worked out the CI issue. This fixes that.

It would be best, of course, to not require that additional step. I'm not currently aware of a solution, though, and think we should update the documentation until one can be found.