fix: Images do not appear in data cards (PT-185149154)
PT Story: https://www.pivotaltracker.com/story/show/185149154

In the context of authoring, `db.stores` is undefined. This change adds a check for that and in the case `db.stores` is undefined, skips the attempt to store an image when it's added to a data card. (The original image URL will be saved instead.)