move tab state into UI store
This replaces the FocusDocumentTracker with the UI store state
The tabs now react to this UI store state, so the commenting system
can update the UI store state and the tabs will change.

The SectionDocumentOrBrowser was split up so now the right side
of CLUE uses a DocumentBrowser instead of reusing
SectionDocumentOrBrowser. There is a lot of overlap though,
so a next step is to extract the common parts of the two.