fix: Renaming Table attrs not showing in linked XYPlot (PT-185317045)
PT Story: https://www.pivotaltracker.com/story/show/185317045

Move the `defaultAxisLabels` check within the `useClickHereCue` check in AttributeLabel so the default labels are only shown when there is no linked data source. This allows the axis label values to be updated when the associated attribute names are changed in a linked data source.