FIX: publishing images added via a datacard tile
these datacard images are stored in the data set, so the
parseDocumentContent function needed to be updated to handle this.

It would be better to make this more generic so we don't have
to update with new tiles or shared models. However there is
still support to avoid publishing multiple images from legacy content.
So until we migrate the legacy content (or give up on the optimization)
we can't make this fully generic.

I also tried to update the README to help others get the tests running.