replace the table unlink handler
This change might affect more things than just DataFlow unlinking.
Previous to this change I'm not sure how adding a new table worked.
I suspect it is using the importedDataSet somehow to provide a default
dataset. If that is true, we might be able to get rid of that code now.

I did check dragging a table from one document to another and that is
still working.