Allow changing arrowhead type of existing vectors.
When changing tools the selection is cleared.
So we can not set the 'vector' button as active when
you just open the palette since that makes it impossible
to select a line and change it into an arrow. So we need
to wait until the user has chosen a vector type, then
update any selected items, and only then set the vector tool active.