FIX scroll jittering
The old code would trigger the reaction again and again.
I didn't find out exactly why, but it was partly because the
`data` of reaction was an object that was being recreated
each time something changed in the UI Store. This was fixed
by having the data be a simple string. It could also have been
fixed by passing options to the reaction of `{equals: comparer.shallow}`

PT-185826033