use ResizePanelDivider in all cases
previously the WorkspaceExpander or ResourcesExpander
was used outside of ResizePanelDivider when one side
was collapsed.
Now the styles of the expanders just change depending
on the position of the divider.
This makes animation of the divider smoother.

This commit also simplifies the Workspace since all the
info the ResizePanelDivider needs in the stores.

Note: some Cypress tests are broken because they expect the nav-tab-panel to not exist in the dom.
It is a goal to full remove the nav-tab-panel from the dom
after animating it. So when that is done the test should
start working again.