WIP: this adds the import/order eslint plugin
it isn't working properly though because it doesn't put the scss/sass
after the other stuff.

However the pattern seems good, we combine this with,
the eslint fix on save, and this way we have
a lot of flexibility on import ordering, and
it will get shown in the CI builds if a developer
doesn't have their IDE configured correctly.