these tests are failing locally and on GitHub
I'm not sure if they always fail.
When I ran them locally, the intented comment was deleted, however
there were additional comments probably from before that still existed.
So the test looking for the comments failed.

As far as I know comments are not cleared by "QA Clear", so the test
will have to deal with that. One suggestion is to add a random unique
id to the comment text, this way the two tests can look for the
specific comment that that was added before. Across retries this id
might change though, so these two tests would not be retry-able.
They current depend on each other so that seems OK.
You could add a option to the individual test to indicate they
shouldn't be retried (I think).