use MobX object instead of just observable.map
MobX was printing warnings in some cases when the map was modified
outside of an action.
The MobX object encapsulates this better and makes sure the map is only
modified inside of an action.

You need to be running CLUE locally with code previous to this.
Then open the standalone doc editor and add a table and a graph.