withoutUndo with opt out
this fixes warnings that happened when the resizing of a row
was being triggered by parent action.

In these cases it seems to make sense to just go ahead and record
the resizing. The reason we didn't record it before because
it could happen on document load and would result in
a weird undo history.