tests of mobx scheduler and mst flow
The mobx flow should work the same.
We are not using these features yet.
The flow test is useful for some planned changes to the ImageMap.
The scheduler test is useful for some planned changes to
updateAfterSharedModelChanges, so it can use MobX tracking and
still be paused during undo/redo.