fix importing tiles with shared model refs
There were 2 issues with old approach:
1. the shared models were added to the document after the tiles
2. the tiles were created independently from the document

The core problem with both issues is that references in the tiles to shared model objects could not be resolved.  The solution for #1 was to first add the content of the shared models to the doc, then the tiles, and then the tile references.  The solution for #2 required refactoring the import code so the tiles would be created within the content by passing the snapshot of the tile to `tileMap.put`.

The refactored import code replaces some deeply nested calls to add* methods in BaseDocumentContent. This add* methods could not be removed because they are still be used in other places. This does mean there is some duplication with those add* methods, but I think the new approach is easier to reason about. It also unifies some parts that were duplicated before. I did the refactoring by "exploding" all of the nested calls and then refactored the result to remove duplication and unnecessary code.

A couple of tests had to be adjusted because the sequence of unique id creation
has changed slightly.

(cherry picked from commit 1a3abb6c5fa35a585ec30b98f68a32c99972e64a)