add cmsEditorBase url param
To make this work I had to relax the origin checks in cms-editor. This seems safe to do, this editor is basically made to be embedded on any page in an iframe.

I also added a configuration for the serve package. By default it redirects *.html to just *. And when it does this it loses the url parameters. This breaks the embedding of cms-editor.html used by the iframe-control. Testing with serve is useful to make sure the admin.html file is working properly without having to deploy it.