fix cypres copy tiles test
I believe the main issue was that a chrome update changed how the temporary scroll bar interacts with the elementFromPoint.
In our tests the temporary scrollbar was covering up the drag handles which would then break the diagram tile because it only allows dragging from the drag handle.

The fix for this is to disable cypress automatic scrolling when triggering events.

In addition to this I also made the drag events mimic how real events work a bit more closely. And I also made a reusable dragTile function since it was now pretty complicated.