make image url match more strict
This approach is still fragile. It would be better to only allow localAssets in specific cases. And then we could require the image urls have a protocol which will make the pattern even less likely to be falsely identified.

Also there are two different tests for image urls. One is in the image map. The other is in data-types.ts. The test in data-types.ts is much more strict, but it likely prevent cases where an image is dragged from another page into CLUE.