add plot control and add it to the generator block
The plot button is not inline with the with value control.
That is handled by the node styling.

This also adds the 'node' class to each node which then brings in more of our styling.
It also fixes an issue with multiple nodes being created for the same MST model.