optimize Dataflow tick
This combines all of the changes made during the onTick and data methods of each node into a single MST action.

It was necessary to fork our own version of the rete dataflow engine which runs synchronously to make this work properly.

The multiple changes per tick caused several negative effects:
- multiple history events were created each tick. The history code gets slower the more history events there are, so this caused a slow down.
- multiple react renders were triggered one for each MST change was completed.  This caused a lot of overhead for each change.
- more storage was needed since there is a wrapper around each of the history events
- the snapshot changes triggered network updates to store the document content.