improve exported content loading
- This changes the loading to manipulate the snapshot instead of a temporary MST document
- It also updates the general document loading to run `typecheck` if there is an error while
loading the document. This results in better error messages when the document is corrupt.
- It updates the dc-general.test.ts to check the structure of the document using a
representation of all of the rows of the document instead of a collection of methods.

A side effect of the loading change is that fewer ids are created, so the tests which compare
snapshots with ids had to be updated.