Add support for downloading offering info
If the offerings are from the portal it gets the info from the portal.
If the offerings are demo offerings, it computes it from the offering id.

This also adds a function to resolve files relative to the script root.

The package-lock.json is not updated because of ongoing work.