Decided to just leave in the Placeholder tile count
One downside to this is it will make it a little harder to identify empty documents.  However firestore doesn't have a query for an empty array.  If we want to identify empty documents we should include a tileCount field in the metadata and placeholders should not be included in this tileCount.