fix reading of comments when there are none
There are cases in the code where a teacher tries to read comments on a document that doesn't have a metadata document. This was causing an Firestore access error.

Also this fixes an issue where the built in exists function from Firestore rules was being overwritten by our own function.