local storage dev root id
This puts a random id in local storage and uses it for the Firestore and Realtime database root ids.
This approach is currently broken though, because the Firebase functions have not been updated to use this same root id. The functions continue to use the Firebase user id instead.

See this PR for more info: https://github.com/concord-consortium/collaborative-learning/pull/2403