make sure documents have problem info
This removes a duplicate creation of the metadata documents from the tree-manager. This duplication could create metadata documents for problem docs that didn't have problem, investigation, and unit info.

It also add ths project, investigation, and unit info to the DocumentModels for problem, planning, and published document models which are created by the db listeners. This info is used by the title code now so it has to be set on the document models all of the time now.