look up users instead of copying full name and initials
This will allow us to update the users later and have the group views automatically update without having to also update the group user objects.

This also removes the option to skip unknown students. They are part of the group so we should show them even if we don't know who they are yet.