feat: Add requests for researcher portal and Firebase JWTs [PT-188754102]
This adds a "researcher" url param to CLUE that, when present, is passed to the requests for the portal and Firebase JWTs.  When the researcher param is set to "true" the portal JWT will mint a JWT with the user_type set to "researcher" for any user.  When it is set to "true" for the Firebase JWT the user must be a researcher for the provided class_hash param.

The portal store was also updated to allow for researcher authentication.  However, once authenticated the user won't be able to currently read any data from Firestore.  There is a follow on PT story for updating the Firestore rules to allow researchers access to the data.