fix: Add missing researcher and target_user_id params to portal requests [PT-188754125]
A utility function called maybeAddResearcherParam was added to add the researcher=true query param to portal api urls if it is set in the CLUE query parameters.  This ensures the portal anonymizes the student info even if the user has higher privileges than a researcher.

The target_user_id is also passed to the portal JWT and Firebase JWT calles if the targetUserId parameter is set in the CLUE query parameters.  This is needed for researchers to check if they have access to the user's information.