feat: Add My Resources Toolbar [CLUE-1]
This adds a separate toolbar to all document views on the left of the workspace.

The toolbar has options set by the caller to show individual buttons.

For now the edit button can be shown and a placeholder for the playback button
is in the code but forced to be hidden until we get to that story.  The
playback option was added just to test the filtering.