fix: Added selection and ordering of all linked tiles when copying [CLUE-3]
The copy spec now determines all the tiles linked to the selected tiles, including transitive links and then orders all the files in document order before creating the copy spec.  This allows the user for example to select a table tile that is at the "end" of three tables linked together and all three tables (and annotations) will be copied in the same order they appear in the document.

NOTE: this probably should also be done for drag/drop and the code is setup for that but it should be done in another story for better QA.