feat: Add standalone app page [CLUE-72]
This adds a new top level page called "standalone".  For now standalone mode shows the curriculum on the left and a TBD panel on the right.

When the login UI is added in the next PR the "STANDALONE FIXME" comments will be addressed, probably with additional or modified model state.