feat: Add check for unit parameter in standalone [CLUE-90]
This adds a check for the unit parameter in the standalone mode.

If the unit parameter is not present, an error alert will be displayed to the user using a custom model error content.

The ability to show custom modal error content was also added in this commit along with the following changes to modals:

- Removed the fixed height of the modal content
- Removed the bottom button area when no buttons are defined
- Updated the styling to allow for the bottom button area to be hidden
- Added a default error title of "Uh Oh!" to the error alert