feat: Add unit navigation for students [CLUE-76]
This allows students to navigate between units in the app.  Before only teachers could navigate between units.  This change requires the portal API to be updated to allow students to request their own offerings.

This also updates unrelated code in the standalone auth component to change a variable name and update a comment based on an earlier PR comment.