feat: Add shareable link to teacher dropdown [CLUE-91]
This adds the functionality to share a link to the current standalone app instance in the teacher dropdown.

This does not update the dropdown styling to match the new spec.  That will be done in a follow on PR as there are a couple of styling issues that need to be resolved first and the current header styling is a bit of a mess.

This commit also updates the version output in the teacher app header to match the student app header and adds a gap between the version and teacher dropdown.