fix: Remove heights from custom modals [CLUE-113]
This removes all the explicit heights from the custom modals as they are not needed anymore as the custom modal was updated in CLUE-90 to add default padding to the modal content which accomplishes the same effect as the height setting but without having to determine and set the height.

This also updates the select styling in the link tile dialog to match the spec'ed padding.