fix: CLUE registration crash due to sign_up_path to long on the portal [CLUE-135]
This fix shortens the url sent to the portal during login/registration.  This is url saved in the database and is used to redirect after registration is successful on the portal.  The database has a current limit of 255 characters for the url and the url used by CLUE just exceeded that.

The url we were sending had a long authDomain url, it has been replaced by a much shorter sentinel value.

A more future proof solution would be to extend the size of the sign_up_path column in the portal from a varchar(255) to varchar(512) but that would require a migration over the nearly 2 million existing user accounts.