fix: Only show students the problem switcher in standalone mode [CLUE-143]
This updates the app header to only show the problem switcher to students in standalone mode.

Before standalone mode was added the students had no access to the switcher.  While standalone mode was being developed it was changed to show for all students, however we only added support to switching classes for students while they were in standalone mode so now the original non-dropdown is shown unless the student is in standalone mode.