fix: Update metadata watcher to watch all unit code variants [CLUE-205]
This updates the sorted-documents.ts#watchFirestoreMetaDataDocs function to use a new CurriculumConfig#getUnitCodeVariants function that is passed the unit and returns an array that includes that unit plus and old unit names that have been renamed to the passed unit.  This ensures that metadata documents that use older unit names are found.