feat: Annotate changes to copied tiles [CLUE-235]
This adds a createdHash and updatedHash to all tiles.  The createdHash is set at object create time and the updatedHash is set whenever the content changes.  The hash is calculated using the tile's exportJson() method and each tile can optionally change what is exported by checking the forHash option in the tile export options to filter out non-content changes, like zoom for the geometry content.