build: Update default CMS port to 8082 and add README
This adds a README.md to the cms folder that points to the existing cms.md folder and also includes other info.

This also sets the default port of the cms to 8082 to reduce the complexity of the order of startup of the app, the git proxy and the cms.