fix plotting of variables
The issue was that the variables need a unitsManager to compute their value now. The plotting code creates a copy of the variables so the x variable value can be scanned across the domain of the graph. The SharedVariables model provides the unitsManager, but it was not part of the copy previously.

Additionally this adds some tests for this behavior as well as the usage of the graph x axis domain when the x variable doesn't have a value.