CLUE-244 try to fix the table tile specs
Note that really it'd be better if the TableToolTile object was split up and the methods that require a table tile would be in a class that was created with a table tile selector. This way all of these method would require the user to specify which table tile they are interacting with.