fix: Straight arrows scale with tiles [CLUE-247]
Straight arrows now maintain their relative position on source tiles when those tiles are resized.

This is achieved by storing a normalized offset based on the source tile's dimensions.

NOTE: as part of this change the trigger for re-rendering the annotation layer was modified
from when the row order changed to when the tile order changed across all rows.  This is to ensure
that the annotation layer is re-rendered when tiles are moved between rows.