CLUE-249 Fix document writing issues, and ignore recent documents
The usage of the BulkWriter was incorrect.
1. It is not safe to wait for bulkWriter create or update calls. These will block until the write is flushed, closed, or additional operations are added to fill up a "batch". So instead of waiting for them directly we wait for a flush call.
2. If the bulk writer operation throws an error and it isn't caught node.js will exit the process. So all places that do bulk writer operations have catch methods added to them to prevent node from exiting. This overlaps with the onWriteError code, but it doesn't seem worth cleaning that up now.

Additionally documents created recently or history entries added recently now cause the script to not migrate those documents. This is to try to reduce problems with running the script while users are updating documents. If they load a document before the script processes it, and then the script processes it while they have the document loaded, the history entries for their new work will get lost and they might see Firestore errors when the history entries are trying to be written.