Authoring Spike [CLUE-287]
This adds the start of a new authoring interface for CLUE, with a left-hand navigation tree based on the unit configuration.

Currently only the content files are "editable", using a tweaked version of the document editor used in the current `cms` folder for Decap.  Edits are currently not saved.

The git backend is faked for now with the branch names as constants and all data loaded from a generated `data.ts` file using the current CLUE curriculum repo using a script included in this commit.

Follow on work, not included here, include:

1. Implementing the new authoring UI requests (without saving)
2. Adding real git support, including commits, branches, and pushing to remote.  We can remove the `data.ts` file at this point.
   We should be able to reuse a lot of the existing Decap code as they have split the backend into multiple packages using a monorepo.
   This work will probably be done as a Firebase function as some of the new authoring requests require simultaneous changes to multiple files whereas Decap focused on single file commits.
3. Adding real saving of edits to content files.