feat: Add auth to authoring [CLUE-292]
This adds Firebase email/password authentication to the authoring tool.

It includes sign-in, sign-out, and user state management.  A custom hook is used as the FirebaseUI package is not compatible with Firebase 8.

This may be extended in the future to support Gmail OAuth to allow CC Google Workspace users to sign in with their Google accounts and automatically be granted access.

To enable local testing, the Firebase Auth emulator is now supported via the `?auth=emulator` URL parameter.

NOTE: this also centralizes Firebase initialization to avoid duplicating code in the main app.