feat: Add authoring backend [CLUE-292]
This commit introduces a new backend API for authoring.

The authoring API is an Express app, and it provides endpoints for managing content,
branches, and units. The backend interacts with GitHub to pull content
and stores metadata in Firebase Realtime Database.

Each of the endpoints requires a valid Firebase ID token for authentication and, for now,
only allows access to users who have validated emails within the concord.org domain.

Until an admin interface is built, units need to be manually pulled by uncommenting the api
call to `/pullUnit` in `src/authoring/app.tsx` with the appropriate parameters set.

Final note: the GitHub API is being accessed in this initial commit without using an auth token.
In the follow-on Git push story, support for using a GitHub token will be added.