feat: Use Firebase ref for pulled files in useCurriculum hook [CLUE-289]
Changed useCurriculum hook to use a Firebase ref to listen for changes to pulled files in real-time,
instead of fetching them via the API.  This allows for more immediate updates when files are modified
by the user or other users (such as a new image being uploaded via the media library).

The pulled files contain metadata about files in the unit, such as their sha and the type of file.
Writes to this data still go through the API.