fix: Add special admin email exception
This commit introduces a special case for Doug's old zoopdoop.com email address
that seems to be stuck in Firebase's auth database even though it is no longer
used on GitHub.

This also updates the temp admin api calls in the app component to wait for
the auth state to be fully initialized before making calls.