feat: Add fake auth user for development
This commit introduces a fake user for authentication during development.

This allows developers to test features that require user authentication
without needing to log in with a real account as GitHub only allows 10 OAuth logins
before it re-prompts the user to authenticate in the popup and then disables the
authentication button for several seconds when the logins are within a small
time range.

This also cleans up some vestigial auth state that was left over from the
initial use-curriculum implementation.