feat: Add authoring mode preview [CLUE-298]
This commit introduces a new authoring mode preview feature that allows users
to see real-time updates of their content changes.

A new url param, `authoringBranch`, has been added to enable this feature.
When present, the runtime will fetch raw content from the authoring branch
instead of the published branch, allowing authors to preview their edits.
This also sets the base curriculum url to use the authoring branch so that
images are also loaded from the raw authoring branch url.

This same param is also passed to the content iframe in the authoring workspace
so that the iframe can also load content from the authoring branch.

A new hook, `useAuthoringPreview`, has been added to encapsulate
the logic for determining if the preview mode is active.  It is used both
in authoring to launch preview windows and in the runtime to check if
the authoring window is still open.  The authoring system uses this hook
to reload all previews when content is changed.

Finally, the curriculum tab authoring compoent has been updated to fix
a bug where the title was not being updated correctly.