feat: Add teacher guides to left nav
This adds is teacher guides to the left nav if they exist.

The use-curriculum hook was modified to return the teacher guide config
if it exists.

This also fixes some layout issues with the left nav so that it uses
the full height of the viewport and the bottom section is always at
the bottom with the tree view properly scrollable.