feat: Add authoring admin UI [CLUE-288]
This adds an admin UI to the authoring tool that allows users to view and manage
branches and units in the authoring environment.  The admin UI is wrapped in
a new Modal component that is also now used by the media library and will be
used by the commit UI in a future PR.

This also adds `main` branch protection to the authoring API to prevent
accidental modifications to the main branch.