feat: Add commit UI [CLUE-288]
This adds a "Commit" button in the authoring header that opens a modal dialog
to show the list of updated files and allow committing them to the current
branch.

This also adds an "Updates" button to the admin table which shows the same
list of updated files.

Note: the actual commit operation is not yet implemented on the backend.