fix: Add deployment of authoring iframe [CLUE-307]
This fixes the issue were the authoring iframe was not being deployed as part of the production release.  To ensure the iframe deployment isn't accidentally removed when/if the cms is removed a separate iframe.html is written out in a subfolder.  This also makes deployment easier as the html filename will be index.html instead of iframe.html.