CLUE-252 summarize dataflow programs to graphviz
LLMs have ingested lots of graphviz descriptions of nodes and links, so they are good at understanding that format.

The approach used here is not easy to maintain. It is duplicating a lot of code. For example the mapping of node operators to formulas.

This summarization is sometimes run in firebase functions, so we can't depend on all of CLUE. However we could extract some of this code into shared so it is used by both the runtime and the server side summarizer.